// types/componentTypes.ts
// for navbar

export interface NavlinksProps{
    title: string;
    href: string;
}



// for Skillsbar creating Type 
export type SkillsbarProps = {
    title: string;
    level: number;
    

} 




// For Project Card 
export interface ProjectCardProps{
    title: string;
    description: string;
    image: string;
    url: string;
};